+++ src/components/bookmark/BookmarkStatus.tsx
@@ -4,7 +4,7 @@ import { FaStar } from 'react-icons/fa';
 import { FaRegTrashAlt } from 'react-icons/fa';
 import { formatNumber } from '@/utils/formatNumber';
 import { getTimeAgo } from '@/utils/getTimeAgo';
-import { VideoStatsProps } from '@/components/video/VideoStats';
+import { VideoStatsProps } from '@/types/video';
 
 const BookmarkStatus = ({
   created_at,
@@ -13,7 +13,7 @@ const BookmarkStatus = ({
   is_bookmarked,
 }: VideoStatsProps) => {
   return (
-    <div className="text-gray-600 flex items-center gap-5">
+    <div className="text-gray-600 mt-1 flex justify-between">
       <div className="flex flex-grow items-center font-bold">
         <div className="mr-2 flex items-center">
           <IoHeartOutline size={16} className="mr-1" />
@@ -23,12 +23,10 @@ const BookmarkStatus = ({
           <VscComment size={16} className="mr-1" />
           <span className="mr-3 text-xs">{formatNumber(comments.length)}</span>
         </div>
+        <time className="flex flex-grow items-center">
+          <span className="mr-3 text-xs">{getTimeAgo(created_at)}</span>
+        </time>
       </div>
-
-      <time className="flex flex-grow items-center">
-        <span className="mr-3 text-xs">{getTimeAgo(created_at)}</span>
-      </time>
-
       <div className="flex flex-grow items-center">
         <FaRegTrashAlt size={16} className="mr-1" />
         <FaStar size={16} className={is_bookmarked ? 'text-primary' : ''} />
