+++ src/components/video/VideoList.tsx
@@ -1,9 +1,5 @@
 import VideoItem from './VideoItem';
-import { VideoProps } from '@/types/video';
-
-export interface VideoListProps {
-  videos: VideoProps[];
-}
+import { VideoListProps } from '@/types/video';
 
 const VideoList = ({ videos }: VideoListProps) => {
   return (
